// src/api/generated/getMovieResponse.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Cinema Platform API
 * API Gateway for Cinema Platform microservices
 * OpenAPI spec version: 1.0
 */

export interface GetMovieResponse {
  id: string;
  title: string;
  slug?: string;
  description: string;
  poster: string;
  banner?: string;
  duration: number;
  ratingAge?: number;
  country?: string;
  releaseDate?: string;
  releaseYear?: number;
  categoryId?: string;
}