// src/api/generated/paginatedMoviesResponse.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Cinema Platform API
 * API Gateway for Cinema Platform microservices
 * OpenAPI spec version: 1.0
 */
import type { GetMoviesResponse } from './getMoviesResponse';

export interface PaginatedMoviesResponse {
  data: GetMoviesResponse[];
  page: number;
  limit: number;
  total: number;
  totalPages: number;
}